Use uint8_t for conversion points

+++ engine/engine/Conversion.hpp
@@ -4,14 +4,15 @@ class Hero;
 #include "engine/DungeonSetup.hpp"
 #include "engine/Monster.hpp"
 
+#include <cstdint>
 #include <functional>
 
 class Conversion
 {
 public:
   Conversion(const DungeonSetup&);
-  int getPoints() const;
-  int getThreshold() const;
+  uint8_t getPoints() const;
+  uint8_t getThreshold() const;
 
   // Add points, returns true if threshold was reached
   [[nodiscard]] bool addPoints(unsigned pointsAdded);
@@ -19,7 +20,7 @@ public:
   void applyBonus(Hero& hero, Monsters& allMonsters);
 
 private:
-  int points;
-  int threshold;
+  uint8_t points;
+  uint8_t threshold;
   std::function<void(Hero&, Monsters&)> bonus;
 };
